feat: add initial implementation of the `.bik` video decoder
Consists of a stream reader, header parser, demuxer, and audio/video decoders. Version 1 of the
`.bik` video format is supported, except for revision `b`. DCT and DFT compressed audio are both
supported. Each outputted frame combines a single video frame with a chunk of decoded audio.

Use tsdown to build the decoder library, pnpm for package management and node as the runtime.

Add initial README documentation, license and an empty changelog.

Initialize CI with a config for Dependabot and workflows for formatting, linting, type-checking,
releasing and publishing.

Refs: #1, #2