feat: add demo with an example video player and documentation
Demo consists of a website with library documentation and an example video
player for `.bik` files. A pre-defined list of downloadable videos is
provided, or alternatively a video can be uploaded to the webapp.

Demo is built using Astro and Starlight, and uses a customized version of the
Starlight Rapide theme.

ci: deploy the demo site when releasing instead of just the documentation

perf: optimize build settings to reduce the minified size of the decoder

docs: improve API documentation